[#13] 기본배송지 등록 기능 구현 (#184)
* [Feat] 기본배송지 등록 기능 구현 (#13)

* [Fix] DB작업 트랜잭션 처리 (#13)

* [Refactor] 기본배송지 반환 함수 로직 변경 (#13)

+++ backend/src/main/java/org/example/backend/domain/delivery/controller/DeliveryController.java
@@ -39,10 +39,26 @@ public class DeliveryController {
     public BaseResponse deleteDelivery(
             @PathVariable Long idx
     ){
+        if (idx == null || idx <= 0){
+            throw new InvalidCustomException(BaseResponseStatus.FAIL);
+        }
         if(!deliveryService.isExist(idx)){
         throw new InvalidCustomException(BaseResponseStatus.USER_DELIVERY_REMOVE_FAIL_NO_MATCH_DELIVERY);
         }
         deliveryService.deleteDelivery(idx);
+        return new BaseResponse();
+    }
+
+    @PatchMapping("")
+    public BaseResponse setDefault(
+            @RequestBody DeliveryDto.SetDefaultRequest request,
+            @AuthenticationPrincipal CustomUserDetails userDetails
+    ){
+        if (request.getIdx() == null || request.getIdx() <= 0){
+            throw new InvalidCustomException(BaseResponseStatus.FAIL);
+        }
+        deliveryService.setDefault(request.getIdx(), userDetails.getIdx());
+
         return new BaseResponse();
     }
